+++ admin/confirm-portal-admin.php
@@ -7,7 +7,7 @@
             <h3>Success... <i class="far fa-grin-hearts"></i></h3>
             <p><?php echo $_GET['message']; ?></p>
         </article>
-        <form action="managePortfolio.php?managePortfolio=<?php echo $_GET['managePortfolio']; ?>" method="POST">
+        <form action="<?php echo $_SERVER['HTTP_REFERER']; ?>" method="POST">
             <input type="submit" value="Confirm">
         </form>
     </section>